New Linux Scripts
The rewrite resulted in a revised bash script as well as a new C++ script.
The performance gains of the compiled C++ script in comparison to the interpreted bash script should be not-super-significant, given the move function is essentially the same between both. Ideally, the C++ was meant to be a core script for both Windows and Linux compilation, but unfortunately I have been unsuccessful with modifying this script for proper Windows execution upon compilation (something about a permission error I'll probably look into in the future).